:lipstick: Update backup secret recovery phrase styles & content according to new designs
