+++ src/modules/Settings/components/BackupRecoveryPhrase/components/CopyRecoveryPhrase.styles.js
@@ -1,4 +1,5 @@
-import { themes, colors, boxes } from 'constants/styleGuide';
+import { themes, colors, boxes, fonts } from 'constants/styleGuide';
+import { setColorOpacity } from 'utilities/helpers';
 
 export function getCopyRecoveryPhraseStyles() {
   return {
@@ -36,11 +37,46 @@ export function getCopyRecoveryPhraseStyles() {
         fontWeight: 'bold',
         paddingLeft: 5,
       },
+      recoveryPhraseContainer: {
+        alignItems: 'center',
+        padding: boxes.boxPadding,
+        borderWidth: 1,
+        borderRadius: 8,
+      },
+      recoveryPhrase: {
+        marginTop: 30,
+        marginBottom: 30,
+        fontFamily: fonts.family.recoveryPhraseText,
+        fontSize: 18,
+        lineHeight: 32,
+        textAlign: 'center',
+      },
+      copyContainer: {
+        alignItems: 'center',
+      },
+      copy: {
+        fontFamily: fonts.family.contextBold,
+      },
+      recoveryPhraseTitle: {
+        marginBottom: 24,
+      },
     },
 
     [themes.light]: {
       container: {
         backgroundColor: colors.light.white,
+      },
+      recoveryPhraseContainer: {
+        borderColor: colors.light.mystic,
+      },
+      recoveryPhraseTitle: {
+        color: colors.light.blueGray,
+      },
+      recoveryPhrase: {
+        color: colors.light.maastrichtBlue,
+      },
+      copy: {
+        color: colors.light.ultramarineBlue,
       },
       text: {
         color: colors.light.blueGray,
@@ -51,8 +87,20 @@ export function getCopyRecoveryPhraseStyles() {
       container: {
         backgroundColor: colors.dark.mainBg,
       },
+      recoveryPhraseContainer: {
+        borderColor: setColorOpacity(colors.light.white, 0.15),
+      },
       text: {
         color: colors.dark.ghost,
+      },
+      recoveryPhraseTitle: {
+        color: colors.dark.platinum,
+      },
+      recoveryPhrase: {
+        color: colors.dark.white,
+      },
+      copy: {
+        color: colors.dark.ultramarineBlue,
       },
     },
   };
